avances para el examen

+++ src/main/java/cl/duoc/ferremasapp/dto/PagoDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class PagoDTO {
-    private Integer pedidoId;
+    private Long pedidoId;
     private double monto;
     private String metodoPago; // DEBITO, CREDITO, TRANSFERENCIA
 }
